+++ src/app/_hooks/navigation.ts
@@ -33,7 +33,7 @@ const routes: NavigationItem[] = [
 	{ name: 'Motion', href: '/motion', current: false, icon: Move3D },
 	{ name: 'Visual Calibration (BETA)', href: '/calibration', current: false, icon: Video },
 	{ name: 'Realtime Analysis (BETA)', href: '/analysis', current: false, icon: AreaChart },
-	{ name: 'System Logs', href: '/update-logs', current: false, icon: FileText },
+	{ name: 'System Logs', href: '/logs', current: false, icon: FileText },
 	// { name: 'Boards', href: '/', current: false, icon: Cpu },
 	// { name: 'Toolhead', href: '/toolhead', current: false, icon: ArrowDownOnSquareIcon },
 ];
