+++ src/app/logs/page.tsx
@@ -1,0 +1,10 @@
+import { LogsViewer } from '@/app/logs/_components/logs-viewer';
+import { LogsErrorBoundary } from '@/app/logs/_components/logs-error-boundary';
+
+export default function LogsPage() {
+	return (
+		<LogsErrorBoundary>
+			<LogsViewer />
+		</LogsErrorBoundary>
+	);
+}
