--- src/app/update-logs/_components/update-logs-viewer.tsx
@@ -1,974 +1,0 @@
-'use client';
-
-import React, { useState, Suspense, useMemo, useEffect, useRef } from 'react';
-import { trpc } from '@/utils/trpc';
-import { twMerge } from 'tailwind-merge';
-import { useWindowVirtualizer } from '@tanstack/react-virtual';
-import { Button } from '@/components/common/button';
-import { Spinner } from '@/components/common/spinner';
-import { ErrorMessage } from '@/components/common/error-message';
-import { Badge, badgeBorderColorStyle } from '@/components/common/badge';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
-import {
-	DropdownMenu,
-	DropdownMenuCheckboxItem,
-	DropdownMenuContent,
-	DropdownMenuLabel,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
-import {
-	Command,
-	CommandEmpty,
-	CommandGroup,
-	CommandInput,
-	CommandItem,
-	CommandList,
-	CommandSeparator,
-} from '@/components/ui/command';
-import { Separator } from '@/components/ui/separator';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Label } from '@/components/ui/label';
-import { PlusCircledIcon } from '@radix-ui/react-icons';
-
-import {
-	AlertCircle,
-	AlertTriangle,
-	CheckCircle,
-	Clock,
-	Download,
-	RefreshCw,
-	Trash2,
-	FileText,
-	Eye,
-	EyeOff,
-	FileClock,
-	FileCode,
-	FileJson,
-	MoreVertical,
-	Info,
-	Bug,
-	Zap,
-} from 'lucide-react';
-import { formatBytes } from '@/helpers/util';
-import { Modal } from '@/components/common/modal';
-import { AnimatedContainer } from '@/components/common/animated-container';
-
-interface LogEntry {
-	level: number;
-	time: string;
-	msg: string;
-	source?: string;
-	context?: string;
-	errorCode?: string;
-	pid?: number;
-	hostname?: string;
-}
-
-interface LogSummary {
-	totalEntries: number;
-	errorCount: number;
-	warnCount: number;
-	infoCount: number;
-	debugCount: number;
-	traceCount: number;
-	fatalCount: number;
-	lastUpdate: string | null;
-	duration: string | null;
-	success: boolean;
-	logFileSize: number;
-	logFileExists: boolean;
-}
-
-const LOG_LEVELS: Record<
-	number,
-	{
-		name: string;
-		color: string;
-		bgColor: string;
-		badgeColor: string;
-		icon: React.ComponentType<{ className?: string }>;
-	}
-> = {
-	10: {
-		name: 'TRACE',
-		color: 'text-zinc-600 dark:text-zinc-400',
-		bgColor: 'bg-zinc-50 dark:bg-zinc-400/10',
-		badgeColor: 'gray',
-		icon: Bug,
-	},
-	20: {
-		name: 'DEBUG',
-		color: 'text-cyan-700 dark:text-cyan-400',
-		bgColor: 'bg-cyan-50 dark:bg-cyan-400/10',
-		badgeColor: 'sky',
-		icon: Bug,
-	},
-	30: {
-		name: 'INFO',
-		color: 'text-green-700 dark:text-green-400',
-		bgColor: 'bg-green-50 dark:bg-green-400/10',
-		badgeColor: 'green',
-		icon: Info,
-	},
-	40: {
-		name: 'WARN',
-		color: 'text-yellow-800 dark:text-yellow-500',
-		bgColor: 'bg-yellow-50 dark:bg-yellow-400/10',
-		badgeColor: 'yellow',
-		icon: AlertTriangle,
-	},
-	50: {
-		name: 'ERROR',
-		color: 'text-red-700 dark:text-red-400',
-		bgColor: 'bg-red-50 dark:bg-red-400/10',
-		badgeColor: 'red',
-		icon: AlertCircle,
-	},
-	60: {
-		name: 'FATAL',
-		color: 'text-purple-700 dark:text-purple-400',
-		bgColor: 'bg-purple-50 dark:bg-purple-400/10',
-		badgeColor: 'purple',
-		icon: Zap,
-	},
-};
-
-const LogSummaryHeader: React.FC<{
-	isLoading: boolean;
-	summary: LogSummary;
-	onRefresh: () => void;
-	onClear: () => void;
-}> = ({ isLoading, summary, onRefresh, onClear }) => {
-	const clearMutation = trpc['update-logs'].clear.useMutation({
-		onSuccess: () => {
-			onClear();
-		},
-	});
-
-	const generateMockDataMutation = trpc['update-logs'].generateMockData.useMutation({
-		onSuccess: () => {
-			onRefresh();
-		},
-	});
-
-	return (
-		<header>
-			{/* Heading */}
-			<div className="bg-zinc-700/15 backdrop-blur-sm">
-				<div className="mx-auto flex max-w-7xl flex-col items-start justify-between gap-x-8 gap-y-4 px-4 py-4 @screen-sm:flex-row @screen-sm:items-center @screen-sm:px-6 @screen-lg:px-8">
-					<div>
-						<div className="flex items-center gap-x-3">
-							<div
-								className={twMerge(
-									'inline flex-none rounded-full bg-green-400/10 p-1 text-zinc-400',
-									(isLoading || !summary.logFileExists) && 'bg-zinc-400/10 text-zinc-400',
-									!isLoading &&
-										summary.logFileExists &&
-										summary.lastUpdate &&
-										!summary.success &&
-										'bg-red-400/10 text-red-400',
-									!isLoading &&
-										summary.logFileExists &&
-										summary.lastUpdate &&
-										summary.success &&
-										'bg-lime-400/10 text-lime-400',
-									!isLoading && summary.logFileExists && !summary.lastUpdate && 'bg-blue-400/10 text-blue-400',
-								)}
-							>
-								{isLoading ? <Spinner className="h-4 w-4" noMargin /> : <FileText className="h-4 w-4" />}
-							</div>
-							<h1 className="flex gap-x-3 text-base leading-7">
-								<span className="font-semibold text-white">System Logs</span>
-								<span className="text-zinc-600">/</span>
-								<span className="font-semibold text-white">
-									{!summary.logFileExists
-										? 'No Log File Found'
-										: !summary.lastUpdate
-											? 'Ready'
-											: summary.success
-												? 'Last Update Success'
-												: 'Last Update Failed'}
-								</span>
-							</h1>
-						</div>
-					</div>
-					<div className="flex gap-2">
-						<Button variant="outline" size="default" onClick={onRefresh}>
-							<RefreshCw className="mr-1 h-4 w-4" />
-							Refresh
-						</Button>
-						{process.env.NODE_ENV === 'development' && (
-							<Button
-								variant="outline"
-								size="default"
-								onClick={() => generateMockDataMutation.mutate()}
-								disabled={generateMockDataMutation.isLoading}
-							>
-								<FileText className="mr-1 h-4 w-4" />
-								Generate Test Data
-							</Button>
-						)}
-						<Button
-							variant="outline"
-							size="default"
-							onClick={() => clearMutation.mutate()}
-							disabled={clearMutation.isLoading || !summary.logFileExists}
-						>
-							<Trash2 className="mr-1 h-4 w-4" />
-							Clear
-						</Button>
-					</div>
-				</div>
-			</div>
-
-			{/* Stats */}
-			{summary.logFileExists && (
-				<div className="border-t border-white/5 bg-zinc-700/10 backdrop-blur-sm">
-					<div className="mx-auto grid max-w-7xl grid-cols-1 @screen-sm:grid-cols-2 @screen-lg:grid-cols-4">
-						{/* Status */}
-						<div className="border-white/5 px-4 py-6 @screen-sm:px-6 @screen-lg:px-8">
-							<p className="text-sm font-medium leading-6 text-white">Update Status</p>
-							<div className="mt-2 flex items-center gap-2">
-								{!summary.lastUpdate ? (
-									<Clock className="h-5 w-5 text-blue-400" />
-								) : summary.success ? (
-									<CheckCircle className="h-5 w-5 text-green-400" />
-								) : (
-									<AlertCircle className="h-5 w-5 text-red-400" />
-								)}
-								<span className="text-lg font-semibold text-white">
-									{!summary.lastUpdate ? 'Ready' : summary.success ? 'Success' : 'Failed'}
-								</span>
-							</div>
-							<div className="mt-1 text-sm text-zinc-400">
-								{summary.totalEntries} entries • {formatBytes(summary.logFileSize)}
-							</div>
-						</div>
-
-						{/* Log Levels */}
-						<div className="border-white/5 px-4 py-6 @screen-sm:border-l @screen-sm:px-6 @screen-lg:px-8">
-							<p className="text-sm font-medium leading-6 text-white">Log Levels</p>
-							<div className="mt-2 flex flex-wrap gap-1">
-								{summary.errorCount > 0 && (
-									<Badge color="red" size="sm">
-										Errors: {summary.errorCount}
-									</Badge>
-								)}
-								{summary.fatalCount > 0 && (
-									<Badge color="purple" size="sm">
-										Fatal: {summary.fatalCount}
-									</Badge>
-								)}
-								{summary.warnCount > 0 && (
-									<Badge color="yellow" size="sm">
-										Warnings: {summary.warnCount}
-									</Badge>
-								)}
-								<Badge color="sky" size="sm">
-									Info: {summary.infoCount}
-								</Badge>
-							</div>
-						</div>
-
-						{/* Timing */}
-						<div className="border-white/5 px-4 py-6 @screen-sm:px-6 @screen-lg:px-8 lg:border-l">
-							<p className="text-sm font-medium leading-6 text-white">Last Update</p>
-							<div className="mt-2">
-								{summary.lastUpdate ? (
-									<div className="flex items-center gap-1">
-										<Clock className="h-4 w-4 text-zinc-400" />
-										<span className="text-sm text-white">{new Date(summary.lastUpdate).toLocaleString()}</span>
-									</div>
-								) : (
-									<span className="text-sm text-zinc-400">No updates performed yet</span>
-								)}
-								{summary.duration && <div className="mt-1 text-sm text-zinc-400">Duration: {summary.duration}</div>}
-							</div>
-						</div>
-
-						{/* Actions */}
-						<div className="border-white/5 px-4 py-6 @screen-sm:border-l @screen-sm:px-6 @screen-lg:px-8">
-							<p className="text-sm font-medium leading-6 text-white">Actions</p>
-							<div className="mt-2">
-								<Modal
-									onClick={() => (window.location.href = '/configure/api/debug-zip')}
-									title="This archive may contain sensitive information"
-									wide={true}
-									body="Please inspect the contents of the zip before posting it publically. Make sure you use Moonraker Secrets if configuring moonraker for third party services."
-									content={
-										<AnimatedContainer>
-											<h3 className="mb-1 font-medium tracking-tight">The following files will be zipped</h3>
-											<ul className="grid gap-2 pb-2 text-muted-foreground">
-												<Suspense fallback={<Spinner />}>
-													{trpc.debugFileList.useSuspenseQuery()[0].map((file) => {
-														return (
-															<li key={file.path} className="flex items-center gap-2 text-sm">
-																{(file.name.endsWith('.log') || file.name.split('.').slice(-2)[0] === 'log') && (
-																	<FileClock className="h-4 w-4 flex-shrink-0 text-zinc-100/40" aria-hidden="true" />
-																)}
-																{file.name.endsWith('.cfg') && (
-																	<FileCode className="h-4 w-4 flex-shrink-0 text-zinc-100/40" aria-hidden="true" />
-																)}
-																{file.name.endsWith('.json') && (
-																	<FileJson className="h-4 w-4 flex-shrink-0 text-zinc-100/40" aria-hidden="true" />
-																)}
-																<span>
-																	{file.orgPath.replace('/home/pi', '~')} ({formatBytes(file.size)})
-																</span>
-															</li>
-														);
-													})}
-												</Suspense>
-											</ul>
-										</AnimatedContainer>
-									}
-									buttonLabel="I understand"
-								>
-									<Button variant="outline" size="sm">
-										<Download className="mr-1 h-4 w-4" />
-										Download Debug Info
-									</Button>
-								</Modal>
-							</div>
-						</div>
-					</div>
-				</div>
-			)}
-
-			{/* No file state */}
-			{!summary.logFileExists && (
-				<div className="border-t border-white/5 bg-zinc-700/10 backdrop-blur-sm">
-					<div className="mx-auto max-w-7xl px-4 py-8 text-center @screen-sm:px-6 @screen-lg:px-8">
-						<FileText className="mx-auto mb-4 h-12 w-12 text-blue-400 opacity-50" />
-						<p className="text-white">System ready for updates</p>
-						<p className="text-sm text-zinc-400">No update logs yet - run an update to generate logs</p>
-					</div>
-				</div>
-			)}
-		</header>
-	);
-};
-
-const LogEntryComponent: React.FC<{ entry: LogEntry; showDetails: boolean }> = ({ entry, showDetails }) => {
-	const level = LOG_LEVELS[entry.level] || {
-		name: 'UNKNOWN',
-		color: 'text-zinc-600 dark:text-zinc-400',
-		bgColor: 'bg-zinc-50 dark:bg-zinc-400/10',
-		badgeColor: 'gray',
-	};
-	const timestamp = new Date(entry.time).toLocaleString();
-	const borderStyle = badgeBorderColorStyle({ color: level.badgeColor as any });
-
-	return (
-		<div className={twMerge('min-h-[60px] rounded-lg border p-3', level.bgColor, borderStyle)}>
-			<div className="flex items-start justify-between gap-2">
-				<div className="min-w-0 flex-1">
-					<div className="mb-1 flex items-center justify-between gap-2">
-						<div className="flex items-center gap-2">
-							<Badge color={level.badgeColor as any} size="sm">
-								{level.name}
-							</Badge>
-							{entry.source && (
-								<Badge color="blue" size="sm">
-									{entry.source}
-								</Badge>
-							)}
-							<span className="text-xs text-muted-foreground">{timestamp}</span>
-						</div>
-						{entry.context && showDetails && (
-							<Badge color="gray" size="sm">
-								{entry.context}
-							</Badge>
-						)}
-					</div>
-					<p className={`text-sm ${level.color} break-words`}>{entry.msg}</p>
-					{showDetails && (
-						<div className="mt-2 space-y-1 text-xs text-muted-foreground">
-							{entry.errorCode && (
-								<div>
-									Error Code: <code className="rounded bg-muted px-1 text-muted-foreground">{entry.errorCode}</code>
-								</div>
-							)}
-							{entry.pid && <div>PID: {entry.pid}</div>}
-							{entry.hostname && <div>Host: {entry.hostname}</div>}
-						</div>
-					)}
-				</div>
-			</div>
-		</div>
-	);
-};
-
-// Custom faceted filter component for log filtering
-interface LogFacetedFilterProps {
-	title: string;
-	options: string[];
-	selectedValues: string[];
-	onSelectionChange: (values: string[]) => void;
-	disabled?: boolean;
-}
-
-const LogFacetedFilter: React.FC<LogFacetedFilterProps> = ({
-	title,
-	options,
-	selectedValues,
-	onSelectionChange,
-	disabled = false,
-}) => {
-	const selectedSet = new Set(selectedValues);
-
-	return (
-		<Popover>
-			<PopoverTrigger asChild>
-				<Button
-					variant="outline"
-					size="sm"
-					className={`h-8 border-dashed ${disabled ? 'cursor-not-allowed opacity-50' : ''}`}
-					disabled={disabled}
-				>
-					<PlusCircledIcon className="mr-2 h-4 w-4" />
-					{title}
-					{selectedValues.length > 0 && (
-						<>
-							<Separator orientation="vertical" className="mx-2 h-4" />
-							<Badge color="gray" className="rounded-sm px-1 font-normal lg:hidden">
-								{selectedValues.length}
-							</Badge>
-							<div className="hidden space-x-1 lg:flex">
-								{selectedValues.length > 2 ? (
-									<Badge color="gray" className="rounded-sm px-1 font-normal">
-										{selectedValues.length} selected
-									</Badge>
-								) : (
-									selectedValues.map((value) => (
-										<Badge color="gray" size="sm" key={value} className="rounded-sm px-1 font-normal">
-											{value}
-										</Badge>
-									))
-								)}
-							</div>
-						</>
-					)}
-				</Button>
-			</PopoverTrigger>
-			<PopoverContent className="w-[200px] p-0" align="start">
-				<Command>
-					<CommandInput placeholder={title} />
-					<CommandList>
-						<CommandEmpty>No results found.</CommandEmpty>
-						<CommandGroup>
-							{options.map((option) => {
-								const isSelected = selectedSet.has(option);
-								return (
-									<CommandItem
-										key={option}
-										className="space-x-2"
-										onSelect={() => {
-											const newSelection = isSelected
-												? selectedValues.filter((v) => v !== option)
-												: [...selectedValues, option];
-											onSelectionChange(newSelection);
-										}}
-									>
-										<Checkbox checked={isSelected} />
-										<span>{option}</span>
-									</CommandItem>
-								);
-							})}
-						</CommandGroup>
-						{selectedValues.length > 0 && (
-							<>
-								<CommandSeparator />
-								<CommandGroup>
-									<CommandItem onSelect={() => onSelectionChange([])} className="justify-center text-center">
-										Clear filters
-									</CommandItem>
-								</CommandGroup>
-							</>
-						)}
-					</CommandList>
-				</Command>
-			</PopoverContent>
-		</Popover>
-	);
-};
-
-// Enhanced log level selector with icons and counts
-interface EnhancedLogLevelSelectorProps {
-	logLevel: string;
-	setLogLevel: (level: string) => void;
-	summary?: LogSummary;
-}
-
-const EnhancedLogLevelSelector: React.FC<EnhancedLogLevelSelectorProps> = ({ logLevel, setLogLevel, summary }) => {
-	const logLevelOptions = [
-		{ value: 'trace', label: 'Trace', level: 10 },
-		{ value: 'debug', label: 'Debug', level: 20 },
-		{ value: 'info', label: 'Info', level: 30 },
-		{ value: 'warn', label: 'Warn', level: 40 },
-		{ value: 'error', label: 'Error', level: 50 },
-		{ value: 'fatal', label: 'Fatal', level: 60 },
-	];
-
-	const getCountForLevel = (level: number): number => {
-		if (!summary) return 0;
-		switch (level) {
-			case 10:
-				return summary.traceCount;
-			case 20:
-				return summary.debugCount;
-			case 30:
-				return summary.infoCount;
-			case 40:
-				return summary.warnCount;
-			case 50:
-				return summary.errorCount;
-			case 60:
-				return summary.fatalCount;
-			default:
-				return 0;
-		}
-	};
-
-	return (
-		<Select value={logLevel} onValueChange={setLogLevel}>
-			<SelectTrigger>
-				<SelectValue />
-			</SelectTrigger>
-			<SelectContent>
-				{logLevelOptions.map((option) => {
-					const levelConfig = LOG_LEVELS[option.level];
-					const count = getCountForLevel(option.level);
-					const IconComponent = levelConfig.icon;
-
-					return (
-						<SelectItem key={option.value} value={option.value}>
-							<div className="flex items-center gap-2">
-								<IconComponent className={`h-4 w-4 ${levelConfig.color}`} />
-								<span>{option.label}</span>
-								{count > 0 && (
-									<Badge color={levelConfig.badgeColor as any} size="sm" className="ml-auto">
-										{count}
-									</Badge>
-								)}
-							</div>
-						</SelectItem>
-					);
-				})}
-			</SelectContent>
-		</Select>
-	);
-};
-
-// Virtualized log list component with infinite scrolling
-interface VirtualizedLogListProps {
-	entries: LogEntry[];
-	isLoading: boolean;
-	error: any;
-	hasNextPage: boolean;
-	isFetchingNextPage: boolean;
-	fetchNextPage: () => void;
-	showDetails: boolean;
-	showOnlyErrors: boolean;
-	logLevel: string;
-	setLogLevel: (level: string) => void;
-	selectedContexts: string[];
-	setSelectedContexts: (contexts: string[]) => void;
-	selectedSources: string[];
-	setSelectedSources: (sources: string[]) => void;
-	setShowDetails: (show: boolean) => void;
-	setShowOnlyErrors: (show: boolean) => void;
-	sortDirection: 'desc' | 'asc';
-	setSortDirection: (direction: 'desc' | 'asc') => void;
-	contexts: string[];
-	sources: string[];
-	summary?: LogSummary;
-}
-
-const VirtualizedLogList: React.FC<VirtualizedLogListProps> = ({
-	entries,
-	isLoading,
-	error,
-	hasNextPage,
-	isFetchingNextPage,
-	fetchNextPage,
-	showDetails,
-	showOnlyErrors,
-	logLevel,
-	setLogLevel,
-	selectedContexts,
-	setSelectedContexts,
-	selectedSources,
-	setSelectedSources,
-	setShowDetails,
-	setShowOnlyErrors,
-	sortDirection,
-	setSortDirection,
-	contexts,
-	sources,
-	summary,
-}) => {
-	// Create a ref for the container to calculate scroll margins
-	const containerRef = useRef<HTMLDivElement>(null);
-
-	const virtualizer = useWindowVirtualizer({
-		count: hasNextPage ? entries.length + 1 : entries.length,
-		estimateSize: () => (showDetails ? 120 : 80), // Dynamic estimation based on details visibility
-		overscan: 5,
-		scrollMargin: containerRef.current?.offsetTop ?? 0,
-		paddingEnd: (containerRef.current?.offsetParent as HTMLElement)?.offsetTop ?? 0,
-	});
-
-	// Load more items when scrolling near the end
-	const virtualItems = virtualizer.getVirtualItems();
-
-	useEffect(() => {
-		const [lastItem] = [...virtualItems].reverse();
-
-		if (!lastItem) {
-			return;
-		}
-
-		if (lastItem.index >= entries.length - 1 && hasNextPage && !isFetchingNextPage) {
-			fetchNextPage();
-		}
-	}, [hasNextPage, fetchNextPage, entries.length, isFetchingNextPage, virtualItems]);
-
-	return (
-		<div className="px-4 @screen-sm:px-6 @screen-lg:px-8">
-			<div className="mb-4 grid grid-cols-1 gap-4 rounded-lg border border-border bg-muted/20 p-4 md:grid-cols-2 lg:grid-cols-4">
-				<div className="space-y-2">
-					<Label htmlFor="log-level" className={showOnlyErrors ? 'text-muted-foreground' : ''}>
-						Log Level
-					</Label>
-					<div className={showOnlyErrors ? 'cursor-not-allowed opacity-50' : ''}>
-						<EnhancedLogLevelSelector logLevel={logLevel} setLogLevel={setLogLevel} summary={summary} />
-					</div>
-				</div>
-
-				<div className="space-y-2">
-					<Label className={showOnlyErrors ? 'text-muted-foreground' : ''}>Context Filter</Label>
-					<LogFacetedFilter
-						title="Context"
-						options={contexts}
-						selectedValues={selectedContexts}
-						onSelectionChange={setSelectedContexts}
-						disabled={showOnlyErrors}
-					/>
-				</div>
-
-				<div className="space-y-2">
-					<Label className={showOnlyErrors ? 'text-muted-foreground' : ''}>Source Filter</Label>
-					<LogFacetedFilter
-						title="Source"
-						options={sources}
-						selectedValues={selectedSources}
-						onSelectionChange={setSelectedSources}
-						disabled={showOnlyErrors}
-					/>
-				</div>
-
-				<div className="space-y-2">
-					<Label>View Options</Label>
-					<div className="flex gap-2">
-						<Button
-							variant={showOnlyErrors ? 'primary' : 'outline'}
-							size="default"
-							onClick={() => setShowOnlyErrors(!showOnlyErrors)}
-							className="flex-1 justify-start"
-						>
-							{showOnlyErrors ? (
-								<>
-									<Eye className="mr-2 h-4 w-4" />
-									Show All
-								</>
-							) : (
-								<>
-									<AlertCircle className="mr-2 h-4 w-4" />
-									Errors Only
-								</>
-							)}
-						</Button>
-						<DropdownMenu>
-							<DropdownMenuTrigger asChild>
-								<Button variant="outline" size="default" className="px-3">
-									<MoreVertical className="h-4 w-4" />
-									<span className="sr-only">View options</span>
-								</Button>
-							</DropdownMenuTrigger>
-							<DropdownMenuContent align="end">
-								<DropdownMenuLabel>Sort Order</DropdownMenuLabel>
-								<DropdownMenuCheckboxItem
-									checked={sortDirection === 'desc'}
-									onCheckedChange={() => setSortDirection('desc')}
-								>
-									Newest First
-								</DropdownMenuCheckboxItem>
-								<DropdownMenuCheckboxItem
-									checked={sortDirection === 'asc'}
-									onCheckedChange={() => setSortDirection('asc')}
-								>
-									Oldest First
-								</DropdownMenuCheckboxItem>
-								<DropdownMenuSeparator />
-								<DropdownMenuLabel>Display Options</DropdownMenuLabel>
-								<DropdownMenuCheckboxItem checked={showDetails} onCheckedChange={setShowDetails}>
-									Show Details
-								</DropdownMenuCheckboxItem>
-							</DropdownMenuContent>
-						</DropdownMenu>
-					</div>
-				</div>
-			</div>
-
-			{isLoading && entries.length === 0 ? (
-				<div className="flex items-center justify-center p-8">
-					<Spinner />
-				</div>
-			) : error ? (
-				<ErrorMessage title="Failed to load log entries">{error.message}</ErrorMessage>
-			) : entries.length === 0 ? (
-				<div className="py-8 text-center text-muted-foreground">
-					<FileText className="mx-auto mb-2 h-12 w-12 opacity-50" />
-					<p>No log entries found</p>
-					{showOnlyErrors && <p className="text-sm">No errors or warnings in the logs</p>}
-				</div>
-			) : (
-				<>
-					<div className="mb-2 text-sm text-muted-foreground">
-						Showing {entries.length} entries
-						{showOnlyErrors && ' (errors and warnings only)'}
-						{hasNextPage && ' (scroll for more)'}
-					</div>
-					<div
-						key={`virtualizer-${showDetails}`}
-						ref={containerRef}
-						style={{
-							height: `${virtualizer.getTotalSize()}px`,
-							width: '100%',
-							position: 'relative',
-						}}
-					>
-						{virtualizer.getVirtualItems().map((virtualItem) => {
-							const isLoaderRow = virtualItem.index > entries.length - 1;
-							const entry = entries[virtualItem.index];
-
-							return (
-								<div
-									key={virtualItem.index}
-									data-index={virtualItem.index}
-									ref={virtualizer.measureElement}
-									style={{
-										position: 'absolute',
-										top: 0,
-										left: 0,
-										width: '100%',
-										transform: `translateY(${virtualItem.start - virtualizer.options.scrollMargin}px)`,
-									}}
-								>
-									{isLoaderRow ? (
-										hasNextPage ? (
-											<div className="flex items-center justify-center p-4">
-												<Spinner />
-												<span className="ml-2 text-sm text-muted-foreground">Loading more entries...</span>
-											</div>
-										) : (
-											<div className="flex items-center justify-center p-4 text-sm text-muted-foreground">
-												No more entries to load
-											</div>
-										)
-									) : (
-										<div className="mb-2">
-											<LogEntryComponent entry={entry} showDetails={showDetails} />
-										</div>
-									)}
-								</div>
-							);
-						})}
-					</div>
-				</>
-			)}
-		</div>
-	);
-};
-
-// Hook for managing infinite scroll data
-const useInfiniteLogEntries = (
-	showOnlyErrors: boolean,
-	logLevel: string,
-	selectedContexts: string[],
-	selectedSources: string[],
-	showDetails: boolean,
-	sortDirection: 'desc' | 'asc',
-) => {
-	const entriesQuery = trpc['update-logs'].entriesPaginated.useInfiniteQuery(
-		{
-			level: logLevel as any,
-			context: selectedContexts.length > 0 ? selectedContexts : undefined,
-			source: selectedSources.length > 0 ? selectedSources : undefined,
-			showDetails,
-			limit: 50,
-			sortBy: 'time',
-			sortDirection,
-		},
-		{
-			enabled: !showOnlyErrors,
-			getNextPageParam: (lastPage) => (lastPage.hasNextPage ? lastPage.nextCursor : undefined),
-			keepPreviousData: true,
-			retry: 3,
-			retryDelay: (attemptIndex) => Math.min(1000 * 2 ** attemptIndex, 30000),
-			staleTime: 10000, // 10 seconds
-		},
-	);
-
-	const errorsQuery = trpc['update-logs'].errorsPaginated.useInfiniteQuery(
-		{
-			source: selectedSources.length > 0 ? selectedSources : undefined,
-			showDetails,
-			limit: 50,
-			sortBy: 'time',
-			sortDirection,
-		},
-		{
-			enabled: showOnlyErrors,
-			getNextPageParam: (lastPage) => (lastPage.hasNextPage ? lastPage.nextCursor : undefined),
-			keepPreviousData: true,
-			retry: 3,
-			retryDelay: (attemptIndex) => Math.min(1000 * 2 ** attemptIndex, 30000),
-			staleTime: 10000, // 10 seconds
-		},
-	);
-
-	const currentQuery = showOnlyErrors ? errorsQuery : entriesQuery;
-
-	// Flatten all pages into a single array of entries - no client-side sorting needed
-	const allEntries = useMemo(() => {
-		return currentQuery.data?.pages.flatMap((page) => page.entries) ?? [];
-	}, [currentQuery.data]);
-
-	return {
-		entries: allEntries,
-		isLoading: currentQuery.isLoading,
-		error: currentQuery.error,
-		hasNextPage: currentQuery.hasNextPage,
-		isFetchingNextPage: currentQuery.isFetchingNextPage,
-		fetchNextPage: currentQuery.fetchNextPage,
-		refetch: currentQuery.refetch,
-	};
-};
-
-export const UpdateLogsViewer: React.FC = () => {
-	const [logLevel, setLogLevel] = useState<string>('info');
-	const [selectedContexts, setSelectedContexts] = useState<string[]>([]);
-	const [selectedSources, setSelectedSources] = useState<string[]>([]);
-	const [showDetails, setShowDetails] = useState<boolean>(false);
-	const [showOnlyErrors, setShowOnlyErrors] = useState<boolean>(false);
-	const [sortDirection, setSortDirection] = useState<'desc' | 'asc'>('desc'); // newest first by default
-
-	const summaryQuery = trpc['update-logs'].summary.useQuery(
-		{
-			source: selectedSources.length > 0 ? selectedSources : undefined,
-		},
-		{
-			retry: 3,
-			retryDelay: (attemptIndex) => Math.min(1000 * 2 ** attemptIndex, 30000),
-			keepPreviousData: true,
-			staleTime: 30000, // 30 seconds
-		},
-	);
-
-	const contextsQuery = trpc['update-logs'].contexts.useQuery(
-		{
-			source: selectedSources.length > 0 ? selectedSources : undefined,
-		},
-		{
-			retry: 2,
-			keepPreviousData: true,
-			staleTime: 60000, // 1 minute
-		},
-	);
-
-	const sourcesQuery = trpc['update-logs'].sources.useQuery(undefined, {
-		retry: 2,
-		keepPreviousData: true,
-		staleTime: 60000, // 1 minute
-	});
-
-	const {
-		entries,
-		isLoading: entriesLoading,
-		error: entriesError,
-		hasNextPage,
-		isFetchingNextPage,
-		fetchNextPage,
-		refetch: refetchEntries,
-	} = useInfiniteLogEntries(showOnlyErrors, logLevel, selectedContexts, selectedSources, showDetails, sortDirection);
-
-	const handleRefresh = () => {
-		summaryQuery.refetch();
-		refetchEntries();
-		contextsQuery.refetch();
-		sourcesQuery.refetch();
-	};
-
-	const handleClear = () => {
-		handleRefresh();
-	};
-
-	if (summaryQuery.error) {
-		return <ErrorMessage title="Failed to load system logs">{summaryQuery.error.message}</ErrorMessage>;
-	}
-
-	const summary = summaryQuery.data;
-	const contexts = contextsQuery.data || [];
-	const sources = sourcesQuery.data || [];
-
-	return (
-		<main className="@container">
-			{summary && (
-				<LogSummaryHeader
-					isLoading={summaryQuery.isLoading}
-					summary={summary}
-					onRefresh={handleRefresh}
-					onClear={handleClear}
-				/>
-			)}
-
-			{summary?.logFileExists && (
-				<div className="border-t border-white/10 pt-11">
-					<div className="mx-auto max-w-7xl">
-						<div className="mb-6 px-4 @screen-sm:px-6 @screen-lg:px-8">
-							<h2 className="text-base font-semibold leading-7 text-white">System Logs</h2>
-						</div>
-
-						<VirtualizedLogList
-							entries={entries}
-							isLoading={entriesLoading}
-							error={entriesError}
-							hasNextPage={hasNextPage ?? false}
-							isFetchingNextPage={isFetchingNextPage ?? false}
-							fetchNextPage={fetchNextPage}
-							showDetails={showDetails}
-							showOnlyErrors={showOnlyErrors}
-							logLevel={logLevel}
-							setLogLevel={setLogLevel}
-							selectedContexts={selectedContexts}
-							setSelectedContexts={setSelectedContexts}
-							selectedSources={selectedSources}
-							setSelectedSources={setSelectedSources}
-							setShowDetails={setShowDetails}
-							setShowOnlyErrors={setShowOnlyErrors}
-							sortDirection={sortDirection}
-							setSortDirection={setSortDirection}
-							contexts={contexts}
-							sources={sources}
-							summary={summary}
-						/>
-					</div>
-				</div>
-			)}
-		</main>
-	);
-};
