+++ src/cli/commands.tsx
@@ -17,7 +17,7 @@ import { createSignal } from '@/app/_helpers/signal';
 import { getLogger } from '@/cli/logger';
 import { frontend } from '@/cli/commands/frontend';
 import { postprocessor } from '@/cli/commands/postprocessor';
-import { updateLogs } from '@/cli/commands/update-logs';
+import { addLogCommands } from '@/cli/commands/logs';
 import development from '@/cli/commands/development';
 
 export const program = new commander.Command()
@@ -487,8 +487,8 @@ log
 		$({ verbose: true })`logrotate -f ${log}`;
 	});
 
-// Add update-logs as a subcommand of the log command
-updateLogs(log);
+// Add log subcommands directly to the logs command
+addLogCommands(log);
 
 development(program);
